move logic to auth class
