+++ airbyte-integrations/connectors/source-cart/Dockerfile
@@ -21,5 +21,5 @@ COPY source_cart ./source_cart
 ENV AIRBYTE_ENTRYPOINT "python /airbyte/integration_code/main.py"
 ENTRYPOINT ["python", "/airbyte/integration_code/main.py"]
 
-LABEL io.airbyte.version=0.1.6
+LABEL io.airbyte.version=0.2.0
 LABEL io.airbyte.name=airbyte/source-cart
