+++ airbyte-integrations/connectors/source-cart/acceptance-test-config.yml
@@ -5,23 +5,23 @@ tests:
   spec:
     - spec_path: "source_cart/spec.json"
   connection:
-    - config_path: "secrets/config.json"
+    - config_path: "secrets/config_central_api_router.json"
       status: "succeed"
     - config_path: "integration_tests/invalid_config.json"
       status: "failed"
       timeout_seconds: 180
   discovery:
-    - config_path: "secrets/config.json"
+    - config_path: "secrets/config_central_api_router.json"
   basic_read:
-    - config_path: "secrets/config.json"
+    - config_path: "secrets/config_central_api_router.json"
       configured_catalog_path: "integration_tests/configured_catalog.json"
       timeout_seconds: 1800
   incremental:
-    - config_path: "secrets/config.json"
+    - config_path: "secrets/config_central_api_router.json"
       configured_catalog_path: "integration_tests/configured_catalog.json"
       future_state_path: "integration_tests/abnormal_state.json"
       timeout_seconds: 1800
   full_refresh:
-    - config_path: "secrets/config.json"
+    - config_path: "secrets/config_central_api_router.json"
       configured_catalog_path: "integration_tests/configured_catalog.json"
       timeout_seconds: 1800
