+++ airbyte-integrations/connectors/source-cart/source_cart/streams.py
@@ -2,19 +2,15 @@
 # Copyright (c) 2022 Airbyte, Inc., all rights reserved.
 #
 
-import base64
-import codecs
-import hashlib
-import hmac
+
 import urllib.parse
-from datetime import datetime
 from abc import ABC
+from datetime import datetime
 from typing import Any, Iterable, Mapping, MutableMapping, Optional, Union
 
 import requests
 from airbyte_cdk.sources.streams.http import HttpStream
-
-from .source import AuthMethod, CentralAPIHeaderAuthenticator, CustomHeaderAuthenticator
+from source_cart.source import CentralAPIHeaderAuthenticator, CustomHeaderAuthenticator
 
 
 class CartStream(HttpStream, ABC):
@@ -34,10 +30,7 @@ class CartStream(HttpStream, ABC):
 
     @property
     def url_base(self) -> str:
-        if self.authenticator.auth_method == AuthMethod.CENTRAL_API_ROUTER:
-            return "https://public.americommerce.com/api/v1/"
-        else:
-            return f"https://{self.authenticator.store_name}/api/v1/"
+        return self.authenticator.url_base()
 
     @property
     def data_field(self) -> str:
@@ -60,14 +53,11 @@ class CartStream(HttpStream, ABC):
         Some endpoints or sometimes Cart.com API returns a datetie instead of the float value to wait to next request.
         Also after calculating the float when Cart.com return a datetime using the value directly
         causes Server Error after a few attempts. Because of this was created the `server_backoff` variable to give time
-        to server recover from too many requests. 
+        to server recover from too many requests.
         """
         server_backoff = 3
         retry_after = response.headers.get("Retry-After")
         if retry_after:
-            self.logger.info(f"Backoff for stream name: {self.__class__.__name__}")
-            self.logger.info(response.status_code)
-            self.logger.info(response.headers)
             try:
                 return float(retry_after)
             except ValueError:
@@ -84,10 +74,9 @@ class CartStream(HttpStream, ABC):
 
     def request_headers(self, **kwargs) -> Mapping[str, Any]:
         auth_signature = {}
-        if self.authenticator.auth_method == AuthMethod.CENTRAL_API_ROUTER:
-            params = self.request_params(self, **kwargs)
-            auth_signature = self.generate_auth_signature(self, params)
-        return dict({"Cache-Control": "no-cache", "Content-Type": "application/json"}, **auth_signature)
+        params = self.request_params(self, **kwargs)
+        extra_params = self.authenticator.extra_params(params)
+        return dict({"Cache-Control": "no-cache", "Content-Type": "application/json"}, **extra_params)
 
     def parse_response(self, response: requests.Response, stream_state: Mapping[str, Any], **kwargs) -> Iterable[Mapping]:
         response_json = response.json()
@@ -102,21 +91,6 @@ class CartStream(HttpStream, ABC):
             params.update(next_page_token)
         return params
 
-    def generate_auth_signature(self, params) -> Mapping[str, Any]:
-        """
-        How to build signature:
-        1. build a string concatenated with:
-            request method (uppercase) & request path and query & provisioning user name
-                example: GET&/api/v1/customers&myUser
-        2. Generate HMACSHA256 hash using this string as the input, and the provisioning user secret as the key
-        3. Base64 this hash to be used as the final value in the header
-        """
-        path_with_params = f"{self.path()}&{urllib.parse.encode(params)}"
-        msg = codecs.encode(f"GET&{path_with_params}&{self.authenticator.user_name}")
-        key = codecs.encode(self.authenticator.user_secret)
-        dig = hmac.new(key=key, msg=msg, digestmod=hashlib.sha256).digest()
-        auth_signature = base64.b64encode(dig).decode()
-        return {"X-AC-PUB-Site-ID": self._site_id, "X-AC-PUB-User": self._user_name, "X-AC-PUB-Auth-Signature": auth_signature}
 
 
 class IncrementalCartStream(CartStream, ABC):
